Refactor class attribute to make clear
