+++ nlpiper/core/composition.py
